consolidate project types and remaining components
