--- src/feat/Preview/index.ts
@@ -1,1 +1,0 @@
-export * from './Preview';
