+++ src/feat/Projects/index.ts
@@ -2,3 +2,4 @@ export * from './ProjectViewer';
 export * from './ProjectProvider';
 export * from './hooks/useProjectViewer';
 export * from './components/ProjectItem';
+export * from './types';
