+++ src/feat/Projects/types.ts
@@ -1,0 +1,4 @@
+import type { Doc, Id } from '~/convex/dataModel';
+
+export type ProjectId = Id<'projects'>;
+export type ProjectEntity = Doc<'projects'>;
